requirements.txt
pyserial
numpy
pandas
bokeh
colorcet